+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. For commit 
  - config validation
    - fails when config file contains unknown fields (helps spot typos)
    - some light value validation on certain fields
+   - removed recaptcha -- was disabled and not used before.
 
  **BugFixes**:
  - fix scope reset on restart https://github.com/gtsteffaniak/filebrowser/issues/515